feat:profile and wishlist

+++ migrations/000004_profile.up.sql
@@ -7,7 +7,7 @@ CREATE TABLE "profile"(
     "postal_code" INTEGER,
     "gender" INTEGER,
     "country" VARCHAR(50),
-    "mobile" INTEGER,
+    "mobile" BIGINT,
     "address" VARCHAR(255),
     "user_id" INT REFERENCES "user"("id")
 );
